requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.2
django-common-helpers==0.9.1
django-cors-headers==2.1.0
djangorestframework==3.6.3
gunicorn==19.6.0
idna==2.5
psycopg2==2.7.1
python-decouple==3.0
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
whitenoise==3.3.0
